sixth commit

+++ app/src/main/java/com/logbook/todo/notification/NotificationScheduler.kt
@@ -26,7 +26,7 @@ object NotificationScheduler {
 
                 val pendingIntent = PendingIntent.getBroadcast(
                     context,
-                    task.id.toInt(),
+                    task.id,
                     intent,
                     PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
                 )
